Reverse a false optimization in previous MessageQueue tuning

+++ src/main/java/com/veraxsystems/vxipmi/connection/queue/MessageQueue.java
@@ -92,7 +92,7 @@ public class MessageQueue extends TimerTask {
 	}
 
 	private synchronized void releaseTag(int tag) {
-		reservedTags.remove(tag);
+		reservedTags.remove((Integer)tag);
 	}
 
 	/**
